# api-gateway/src/docs/openapi3_0.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Supermarket API",
    "version": "1.0.0",
    "description": "A Node.js Express API for managing supermarket-related services including authentication, users, products, and purchases. The system is designed with scalability and maintainability in mind, separating concerns into independent services that follow the  Model-View-Controller(MVC) design pattern and communicate through an API Gateway."
  },
  "servers": [
    {
      "url": "http://localhost:3000/supermarket/",
      "description": "Base URL for the supermarket API"
    }
  ],
...
  "paths": {
...
    "/producs": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Retrieves a paginated list of products with optional filters.",
        "operationId": "getFilteredProducts",
        "description": "Returns a list of products based on optional query parameters such as category, stock, price range, limit, and offset. The response includes pagination details.",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Filter products by category.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Categories"
            }
          },
          {
            "name": "stock",
            "in": "query",
            "description": "Filter products with stock greater than this value. Must be a positive number.",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "description": "Filter products with a price greater than or equal to this value. Must be a positive number.",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "description": "Filter products with a price less than or equal to this value. Must be greater than `minPrice`.",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of products to return per page. Defaults to 10.",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of products to skip for pagination. Defaults to 0.",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the paginated list of filtered products.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "products": {
                      "$ref": "#/components/schemas/FilteredProducts"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationParametersError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "products"
        ],
        "summary": "Create a new product",
        "operationId": "createProduct",
        "description": "Creates a new product with the provided details.",
        "security": [
          {
            "AccessToken": []
          },
          {
            "RefreshToken": []
          }
        ],
        "requestBody": {
          "description": "Create a new product in the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product": {
                      "$ref": "#/components/schemas/ProductResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error in request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZodError"
                }
              }
            }
          },
          "401": {
            "description": "Access not authorized (Invalid or expired token provide)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/products/categories": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Retrieves a list of product categories.",
        "operationId": "getCategories",
        "description": "Returns a list of available categories. If no categories are found, a message indicating that no products were found is returned.",
        "responses": {
          "200": {
            "description": "Successfully retrieved the list of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "categories": {
                      "$ref": "#/components/schemas/Categories"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No categories were found",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Categories not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/products/{id}": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Retrieves a product by its ID.",
        "operationId": "getProductById",
        "description": "Returns a single product based on the provided ID. If the ID is invalid or the product is not found, appropriate error messages are returned.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the product (UUID format).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "5760f04b-fd96-408d-adb4-1ab149371b9a"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the product.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product": {
                      "$ref": "#/components/schemas/ProductResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The provided ID does not have a valid format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidIdError"
                }
              }
            }
          },
          "404": {
            "description": "The product with the specified ID was not found.",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductNotFound"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "products"
        ],
        "summary": "Update a product by ID",
        "operationId": "updateProduct",
        "description": "Partially updates a product using its unique UUID identifier.",
        "security": [
          {
            "AccessToken": []
          },
          {
            "RefreshToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the product (UUID format).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "5760f04b-fd96-408d-adb4-1ab149371b9a"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "aplication/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product": {
                      "$ref": "#/components/schemas/ProductResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format or validation error in request body.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/InvalidIdError"
                    },
                    {
                      "$ref": "#/components/schemas/ZodError"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "The product with the specified ID was not found.",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductNotFound"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "products"
        ],
        "summary": "Delete a product by ID",
        "operationId": "deleteProduct",
        "description": "Deletes a product using its unique UUID identifier.",
        "security": [
          {
            "AccessToken": []
          },
          {
            "RefreshToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the product (UUID format).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "5760f04b-fd96-408d-adb4-1ab149371b9a"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted successfully (no content returned)"
          },
          "400": {
            "description": "Invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidIdError"
                }
              }
            }
          },
          "404": {
            "description": "The product with the specified ID was not found.",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductNotFound"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/products/code/{code}": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Get a product by its code.",
        "operationId": "getProductByCode",
        "description": "Retrieves a product using its unique code.",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "description": "The unique code of the product (must be a valid number)",
            "required": true,
            "schema": {
              "type": "number",
              "example": 1639828085202
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product": {
                      "$ref": "#/components/schemas/ProductResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid code provided.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid code provided.The code must be a valid number"
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "InvalidCode"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductNotFound"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Categories": {
        "type": "array",
        "items": {
          "type": "string",
          "enum": [
            "beauty",
            "fragrances",
            "groceries",
            "furniture",
            "home-decoration",
            "kitchen-accessories",
            "laptops",
            "mens-shirts",
            "mens-shoes",
            "mens-watches",
            "mobile-accessories",
            "motorcycle",
            "skin-care",
            "smartphones",
            "sports-accessories",
            "sunglasses",
            "tablets",
            "tops",
            "vehicle",
            "women-bags",
            "women-dresses",
            "women-jewelry",
            "women-shoes",
            "women-watches"
          ]
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "totalRecords": {
            "type": "integer",
            "description": "Total number of records found.",
            "example": 64
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of available pages.",
            "example": 7
          },
          "currentPage": {
            "type": "integer",
            "description": "Current page number.",
            "example": 1
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Indicates if there are more pages available.",
            "example": true
          },
          "hasPreviousPage": {
            "type": "boolean",
            "description": "Indicates if there are previous pages available.",
            "example": false
          }
        }
      },
      "FilteredProducts": {
        "type": "array",
        "description": "List of filtered poducts",
        "items": {
          "$ref": "#/components/schemas/ProductResponse"
        }
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier of the product.",
            "example": "5760f04b-fd96-408d-adb4-1ab149371b9a"
          },
          "name": {
            "type": "string",
            "description": "Name of the product.",
            "example": "Tissue Paper Box"
          },
          "description": {
            "type": "string",
            "description": "Description of the product.",
            "example": "Convenient tissue paper box for everyday use, providing soft and absorbent tissues."
          },
          "category": {
            "type": "string",
            "description": "Category of the product, will be one of the predefined categories",
            "example": "groceries"
          },
          "price_purchase": {
            "type": "number",
            "description": "Purchase price of the product.",
            "example": 2.49
          },
          "price_sale": {
            "type": "number",
            "description": "Sale price of the product.",
            "example": 3.99
          },
          "stock": {
            "type": "integer",
            "description": "Stock quantity of the product.",
            "example": 87
          },
          "threshold": {
            "type": "integer",
            "description": "Threshold for restocking.",
            "example": 5
          },
          "active": {
            "type": "boolean",
            "description": "Indicates if the product is active.",
            "example": true
          },
          "thumbnail": {
            "type": "string",
            "description": "Thumbnail image URL of the product.",
            "example": "https://cdn.dummyjson.com/products/images/groceries/Tissue%20Paper%20Box/thumbnail.png"
          },
          "code": {
            "type": "number",
            "description": "Unique product code.",
            "example": 1639828085202
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp indicating when the user account was created.",
            "example": "2025-03-15T10:30:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp indicating when the user account was last updated.",
            "example": "2025-03-25T11:56:00Z"
          }
        }
      },
      "ProductRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the product. (max 55 characters)",
            "example": "Tissue Paper Box"
          },
          "description": {
            "type": "string",
            "description": "Description of the product. (cannot be empty)",
            "example": "Convenient tissue paper box for everyday use, providing soft and absorbent tissues."
          },
          "category": {
            "$ref": "#/components/schemas/Categories"
          },
          "price_purchase": {
            "type": "number",
            "description": "Purchase price of the product. (must be positive)",
            "example": 2.49
          },
          "price_sale": {
            "type": "number",
            "description": "Sale price of the product. (must be positive and higher than purchase price)",
            "example": 3.99
          },
          "stock": {
            "type": "integer",
            "description": "Stock quantity of the product.(must be a positive integer)",
            "example": 87
          },
          "threshold": {
            "type": "integer",
            "description": "Threshold for restocking. (optional, must be a positive integer)",
            "default": 0
          },
          "active": {
            "type": "boolean",
            "description": "Indicates if the product is active.",
            "default": true
          },
          "thumbnail": {
            "type": "string",
            "description": "Thumbnail image URL of the product. (optional)",
            "example": "https://cdn.dummyjson.com/products/images/groceries/Tissue%20Paper%20Box/thumbnail.png"
          },
          "code": {
            "type": "number",
            "description": "Unique product code. (must be a 13-digit positive number)",
            "example": 1639828085202
          }
        },
        "required": [
          "name",
          "description",
          "category",
          "price_purchase",
          "price_sale",
          "stock",
          "code"
        ]
      },
      "ProductUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the product. (optional, max 55 characters)",
            "example": "Tissue Paper Box"
          },
          "description": {
            "type": "string",
            "description": "Description of the product. (optional, cannot be empty)",
            "example": "Convenient tissue paper box for everyday use, providing soft and absorbent tissues."
          },
          "category": {
            "$ref": "#/components/schemas/Categories"
          },
          "price_sale": {
            "type": "number",
            "description": "Sale price of the product. (optional, must be positive and higher than purchase price)",
            "example": 3.99
          },
          "stock": {
            "type": "integer",
            "description": "Stock quantity of the product.(optional, must be a positive integer)",
            "example": 87
          },
          "threshold": {
            "type": "integer",
            "description": "Threshold for restocking. (optional, must be a positive integer)",
            "default": 0
          },
          "active": {
            "type": "boolean",
            "description": "Indicates if the product is active(optional).",
            "default": true
          },
          "thumbnail": {
            "type": "string",
            "description": "Thumbnail image URL of the product. (optional)",
            "example": "https://cdn.dummyjson.com/products/images/groceries/Tissue%20Paper%20Box/thumbnail.png"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "ValidationParametersError": {
        "type": "array",
        "description": "List of validation errors on parameters",
        "items": {
          "type": "object",
          "properties": {
            "field": {
              "type": "string",
              "description": "Field with validation error",
              "example": "limit"
            },
            "message": {
              "type": "string",
              "description": "Descriptive error message",
              "example": "Expected number, received string"
            }
          }
        }
      },
      "ZodError": {
        "type": "array",
        "description": "List of validation errors",
        "items": {
          "type": "object",
          "properties": {
            "field": {
              "type": "string",
              "description": "Field with validation error",
              "example": "stock"
            },
            "message": {
              "type": "string",
              "description": "Descriptive error message",
              "example": "Expected number, received string"
            }
          }
        }
      },
      "UnauthorizedResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "You are not authorized to perform this action."
          },
          "code": {
            "type": "string",
            "example": "InvalidToken"
          }
        }
      },
      "InvalidIdError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "The provided ID does not have a valid format"
          },
          "errorCode": {
            "type": "string",
            "example": "InvalidId"
          }
        }
      },
      "ProductNotFound": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Product not found"
          }
        }
      },
...
    },
    "securitySchemes": {
      "AccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header",
        "name": "authorization",
        "description": "Access token passed in the Authorization header"
      },
      "RefreshToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "cookie",
        "name": "refresh_token",
        "description": "Refresh token passed as refresh_token=bearer<JWT> in the Cookie header"
      }
    }
  }
}